+++ shared/src/commonMain/kotlin/com/merseyside/partyapp/data/repository/ItemRepositoryImpl.kt
@@ -20,7 +20,7 @@ class ItemRepositoryImpl(
         eventId: Long,
         name: String,
         description: String,
-        price: Long,
+        price: Double,
         payMember: MemberItemInfo,
         membersInfo: List<MemberItemInfo>
     ): Boolean {
