+++ shared/src/commonMain/kotlin/com/merseyside/partyapp/domain/interactor/AddItemInteractor.kt
@@ -27,7 +27,7 @@ class AddItemInteractor : CoroutineUseCase<Boolean, AddItemInteractor.Params>() 
         val eventId: Long,
         val name: String,
         val description: String,
-        val price: Long,
+        val price: Double,
         val payMember: MemberItemInfo,
         val membersInfo: List<MemberItemInfo>
     )
